// packages/deepspace/src/storage/types.ts
export interface RecordData<T = unknown> {
  recordId: string
  data: T
  createdBy: string
  createdAt: string
  updatedAt: string
}